Move redux to utils
