+++ src/features/board/board.slice.ts
@@ -2,7 +2,7 @@ import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import { getBoard } from "src/features/board/board.api";
 import { Board as innerBoard } from "src/types/board.g";
 import { Board as OuterBoard } from "src/features/board/board.api";
-import { RootState } from "src/redux/root";
+import { RootState } from "src/utils/redux/root";
 
 interface BoardState {
   boards: innerBoard[];
