+++ src/features/column/column.slice.ts
@@ -1,7 +1,7 @@
 import { createSlice } from "@reduxjs/toolkit";
 import { Column as innerColumn } from "src/types/column.g";
 import { Column as OuterColumn } from "src/features/board/board.api";
-import { RootState } from "src/redux/root";
+import { RootState } from "src/utils/redux/root";
 import { getBoardThunk } from "src/features/board/board.slice";
 
 interface ColumnState {
