+++ src/features/user/user.slice.ts
@@ -1,5 +1,5 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
-import { RootState } from "src/redux/root";
+import { RootState } from "src/utils/redux/root";
 import { signin, signout } from "src/features/auth/auth.slice";
 import { getUser } from "src/features/user/user.api";
 import { User } from "src/types/user.g";
