+++ src/utils/redux/store.ts
@@ -4,7 +4,7 @@ import {
   ThunkDispatch,
   Action,
 } from "@reduxjs/toolkit";
-import { rootReducer } from "src/redux/root";
+import { rootReducer } from "src/utils/redux/root";
 import { currentEnv } from "src/utils/server-env";
 import { serverEnv } from "src/utils/server-env";
